Posting downloaders on executor
