+++ core/src/main/java/de/danoeh/antennapod/core/service/download/Downloader.java
@@ -17,7 +17,7 @@ public abstract class Downloader implements Callable<Downloader> {
 
     private volatile boolean finished;
 
-    volatile boolean cancelled;
+    public volatile boolean cancelled;
 
     @NonNull
     final DownloadRequest request;
